#27 - moves last common components to ui module and its tests
