+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/creditcardbillscard/CreditCardBillsCard.kt
@@ -61,7 +61,7 @@ import br.com.rbrthmn.R
 import br.com.rbrthmn.home.ui.components.AddItemButton
 import br.com.rbrthmn.home.ui.components.BanksDropdownMenu
 import br.com.rbrthmn.home.ui.components.TotalValueText
-import br.com.rbrthmn.ui.financialcompanion.components.DecimalInputField
+import br.com.rbrthmn.ui.components.DecimalInputField
 import br.com.rbrthmn.ui.financialcompanion.screens.home.components.creditcardbillscard.CreditCardBillsCardContract.Intent
 import org.koin.androidx.compose.koinViewModel
 import java.time.LocalDate
