+++ core/ui/build.gradle.kts
@@ -68,6 +68,8 @@ dependencies {
     implementation(libs.androidx.runtime.android)
     implementation(libs.androidx.material3.android)
     implementation(libs.androidx.compose.ui.tooling.preview)
+    implementation(libs.koin.androidx.compose)
+    implementation(platform(libs.koin.bom))
 
     testImplementation(libs.junit)
