+++ core/ui/src/main/java/br/com/rbrthmn/ui/components/ReservesDropdownMenu.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.components
+package br.com.rbrthmn.ui.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.material.icons.Icons
@@ -36,7 +36,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import br.com.rbrthmn.R
+import br.com.rbrthmn.ui.R
 
 @Composable
 fun ReservesDropdownMenu(
